function/requirements.txt
# Azure Functions
azure-functions>=1.17.0
azure-functions-durable>=1.2.0

# Azure Identity (for managed identity)
azure-identity>=1.15.0

# Microsoft Graph SDK
msgraph-sdk>=1.2.0

# Azure Management
azure-mgmt-authorization>=4.0.0

# Azure Monitor (for Log Analytics ingestion)
azure-monitor-ingestion>=1.0.3

# Utilities
python-dateutil>=2.8.2
